pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
isort = "^5.13.2"
black = "^24.4.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
isort = "^5.13.2"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'


[tool.black]

# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''


[tool.isort]
# максимальная длина строки
line_length = 119










[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
